mmu working fine now yeaaa

+++ kernel/include/mmu/alloc.h
@@ -2,7 +2,7 @@
 #define MMU_ALLOC_H
 
 #include<stdint.h>
-void *ksbrk(uint32_t);
+void *ksbrk(uint16_t);
 void *kmalloc(uint32_t);
 void kfree(void *);
